Add support for alias users
This allows a single user to manage multiple accounts, albeit without
being able to mark items as bought/unbought. Previously, a user making
use of an alias account would inadvertently end up seeing the 'bought'
status for all their requested items (since they are logged in as a
separate user).

Alias users are useful to create lists on behalf of people who are not
users of the system. It is assumed the bought status can then be
managed via the masquerading user's actual account.